Patch on the fish copy_files upgrade
Made sure the copy_files script didn't mind where it
was called from (used to get pwd, now gets the directory
of the script).

On yoshi, the fish -v command was outputting to stderr, so the
grep search for version didn't work. Modified so both stderr
and stdout were piped to grep.