Mario: put case-switch statement in config.fish
When loading the environment variables for different hosts, I used to copy
a file from this repo to the ~/.config/fish directory, and load that in. The
decision was made at copy time. Now i've moved that decision to be made at
loading time. This helps with yoshi and mario who share the same home folder..